pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localwiki"
version = "0.1.0"
description = "Fully local Karpathy-style self-compiling wiki powered by Ollama"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.3",
    "tavily-python>=0.3",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "python-frontmatter>=1.1",
    "python-dotenv>=1.0",
    "streamlit>=1.35",
    "pyvis>=0.3",
    "langgraph>=0.2",
    "langchain-ollama>=0.2",
    "langchain-core>=0.3",
    "httpx>=0.27",
    "markdownify>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
